Stream parsing now works without memoization
Memoization is not a good idea for this application because the memory
requirements are too high. Basically, the memoizer will hold references
to the entire ast. Thus, to prevent double parsing of the ambiguous
constructs in the `dot` language while the ambiguity is worked out the
grammar was left factored to delay the ambiguous choice. Here is an
example of the problem

digraph {
    {x}->a
}

digraph {
    {x}
}

digraph {
    x
}

digraph {
    x;
}

digraph {
    x -> x;
}

Since, a subgraphs and nodes (without attr clauses) can be both a Stmt
and and EdgeReciever it is difficult to know without arbitrary
look-ahead which is it going to be. If a EdgeOp (-- or ->) appears after
the NodeId or SubGraph then one can be sure it is a EdgeReciever. If some other
token appears afterwards then once can be sure it is a Stmt. However,
since a subgraph can have an arbitrary number of tokens in it, no fixed
amount of look-ahead can solve the ambiguity. You have to deal with it
explicitly in the grammar.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>